requirements.txt
pycryptodome~=3.22.0
flask~=3.1.0

python-socketio~=5.13.0
Flask-SocketIO~=5.5.1